topcat: new VariableWindow for displaying global variables
At present the application-wide instance of VariablePanel is
obtained using the static getInstance method.
This is a bit unsatisfactory (precludes multiple application
instances in the same JVM) but there isn't currently a better place
to store an application-wide instance.  ControlWindow is where most
things like that live, but that's a JFrame which means it can't be
referenced in a headless context, which means headless unit tests
would fail.  At some point in the future, find a better place for
application-wide state.