ttools: fix row selection stilts command generation bug
In some circumstances, if a plot had just one layer with all rows
and just one layer with a limited selection of rows, the selection
(icmd='select <expr>') was not getting inserted into the generated
stilts command.

This was because StiltsCommand.createInputTableSettings was leaving
the filter setting command blank in the input table specification
rather than giving it a default (blank) setting, so that
PlotStiltsCommand.extractCommonSettings was factoring the sole
filter setting command out to the global level, thus applying it
to all layers including the one that should have had icmd=''.
Fix it.