ttools: fix x/ycrowd stilts command generation bug
In the corner plot the xcrowd and ycrowd parameters were being
treated incorrectly when generating a stilts command from topcat,
since stilts thought they would be indexed for each cell of the
plot grid, but the topcat GUI wasn't allowing this.
The result was multiple, perhaps conflicting, values for the
xcrowd parameter in the output stilts command.

Instead don't describe these parameters as XyKeyPairs,
which means that the MatrixPlot processing doesn't try to
insert them for each grid cell.