topcat: save session can now store global information
The TopcatCodec interface now allows application-wide state to be
stored alongside table-specific state in a table prepared for
serialization in a saved session file.  Such information is stored
alongside table-specific state in the first saved table only.

At this commit no such global information is stored.

I experimented with putting it in a separate globals table which
is conceptually cleaner, but it raises backward compatibility
problems if older TopcatCodecs come across a table in the session
file that doesn't represent a loadable TopcatModel.