topcat: rework TopcatModel subsetCounts map
The Map keeping track of subset counts now has LabelledCounts
rather than just Longs as values.  This means that with each
count can be associated a string hashing its content - if the
label changes, the subset contents can be expected to have changed.

This facility is not used at this commit, and behaviour of the
application should be unchanged.