topcat: recount subsets on global variable changes
A listener on global variable values is installed into TopcatModel
that checks if any of the subsets it looks after have apparently
been affected by a value change for one of the global variables.
In this case the subset count is invalidated, which may or may not
trigger a recount but at least the invalid count will no longer
be displayed.