topcat: add check for pointless replots
Avoid replot attempts when the window is invisible.
Now we're getting replot requests triggered by GUI actions outside
of the plot window itself (global variable value changes),
we need to make sure they're not happening after the window is
no longer visible to avoid wasting resources.

Mostly this should be addressed by unregistering listeners when
the window is done with, but it doesn't hurt to have a check
in the replot method as well in case that doesn't catch all
instances.