table: fix RowEvaluator BLANK_DECODER problem
When the Decoder class was parameterised a few weeks ago at ba5bae2d2
BLANK_DECODER got typed as Decoder<Void>.  This was sort of the right
thing to do, but it meant that tables read by the CSV and ASCII input
handlers could end up with Void-typed columns if all the values in
the input columns were completely blank.  Typed output handlers
such as VOTable then choked on such StarTable columns (typically
omitting such columns from output tables) since they don't know how
to write a Void typed-column.

There are a few fixes one could imagine for this, but the least
disruptive is to revert to the previous behaviour and type BLANK_DECODER
as a Decoder<String>.