hapi: replace non-ASCII character in source code
In four places, the apostrophe in "Heliophysics Data Application
Programmer’s Interface" was the character U+2019 (RIGHT SINGLE
QUOTATION MARK).  Since this is not 7-bit ASCII it can cause trouble
when building on platforms with a non-UTF-8 encoding (the source
code is UTF-8 encoded).  I don't know how/why that got in there,
but replace it in all cases with U+0027 (APOSTROPHE).