topcat: provide autocomplete in ColumnDataComboBox
When entering text into a ColumnDataComboBox, a popup menu is now
displayed showing you only the columns matching the text that
you have entered so far.  This should help working with tables
having lots of columns.

The battle with Swing to get this working was extremely fiddly,
and the result is not perfect.  The main remaining issue is that
the down/up keys that you might expect to move the selection around
the popup menu once it's visible don't do what they should.
I have tried pretty hard to improve this but so far failed,
though if there is a solution it's probably only a few lines of code.