starjava: fix source code encoding to ASCII
In all the build files for code that is likely to be changed by me,
make sure that javac is invoked with "-encoding ascii".
If the encoding is not specified explicitly in this way,
the compiler uses the platform default encoding, which means
the code might build in some environments (especially locales)
and not others.

The encoding doesn't have to be fixed at ASCII, it could be
any encoding (e.g. UTF-8) as long as the source files in the
package where the compiler uses that encoding are consistent with it.

For now I've chosen ASCII since all the source files in these packages
are currently 7-bit ASCII only, and having it like this enforces that
state of affairs, so if any non-ASCII characters are introduced
into the sources, the build will probably break and draw attention
to that fact.
This in turn means that the source files can be easily copied to
other places (as long as the encoding is ASCII-compatible,
which is nearly everywhere and in particular UTF-8 environments)
and compile without any changes.
But if in the future some pressing need arises for some or all of
these packages to contain non-ASCII characters in the source code,
the "encoding" property in the build files could be changed from
"ascii" to e.g. "UTF-8" instead.

Note this does not apply to SPLAT, which does have some UTF-8
encoded characters.  Probably that should explicitly state UTF-8
encoding in the same way in its build file, but I haven't done that.