ttools: votlint checks char/unicodeChar characters
For all arraysizes of both char and unicodeChar, votlint now checks
each character to see it's in range (7-bit ASCII or UCS-2).
Previously out of range characters were glossed over.

This applies to FIELD and PARAM values.  At this commit no changes
are made to the handling of INFO values, so any Unicode in an
INFO value attribute will pass without report.  Strictly speaking
this should be checked in the same way as for a PARAM with
datatype="char", since VOTable 1.5 sec 4.8 says:

   "The INFO element is a PARAM element restricted to be of type string
    (i.e. datatype='char' and arraysize='*' are implied)."

But there's really no point in that restriction, since unlike PARAMs
INFOs don't need to deal with string arrays and fixed-length
serializations; really INFO should be defined separately from PARAM,
and a VOTable Erratum to that effect is under consideration
at time of writing.

Although it makes these new checks, at this commit it does NOT
actually report them - the FixedLengthCharType.charError method
is a no-op.  This is because the errors it's now checking for,
e.g. UTF-8 in char FIELDs/PARAMs, though technically illegal,
probably did the right thing in existing software contexts.
Moreover, there's no very good fix for such constructs at VOTable 1.5.
And historically, votlint has not reported these misdemeanours.
When VOTable 1.6 becomes available, then using datatype="char" in
VOTable 1.6 is the obvious way to fix such things, so at that
point the error reporting should be unsuppressed.