topcat: fix Switch X/Y button for multi-coord plots
The switch X-Y button in the Plane Plot was only swapping the
contents of the first two coordinate selectors, which is correct
for simple scatter plots, but incorrect for e.g. pair plots
where there are two X and two Y coordinates.  Fix it.

Thanks to Benjamin Steltner for spotting this.