topcat: don't suppress replots before window is posted
Code in PlotPanel prevents replots happening when the window is
invisible to save resources.  However, this should only kick in
after the plot has been made visible once, since preparation may
be necessary before the panel is actually placed in the window
hierarchy.  Add logic to make sure this is the case.

This fixes a serious regression bug; since this replot suppression
was introduced at f42b8bd30 the Time Plot has been broken, resulting
in NullPointerExceptions when you try to show the window.