topcat: fix variable-related session restore bug
If synthetic columns or subsets were dependent on a global variable
other than one of the default ones (var$x, var$i) they were not
getting restored (WARNING: Can't evaluate column ...).

This was because I was calling things out of sequence in
TopcatCodec2.decode; some were called in SwingUtilities.invokeLater
and others weren't.  Actually it seems the whole of that method
is invoked at restore time on the Event Dispatch Thread anyway.