topcat: remove autocomplete on SortOrderSelector
The column name autocomplete doesn't work very well on the
SortOrderSelector; if you enter a column name and then delete
letters from it you get a cascade of popups telling you that
the content is invalid.

This is presumably because too many or the wrong events are being fired
by this component, and a proper fix would be to track that down and
correct it.  But my initial efforts failed, so instead for now
just disable autocomplete for these ColumnDataComboBoxes so that
the problem doesn't arise.  Since the autocomplete functionality
has only just been introduced, this is not withdrawing a previously
available feature.