parquet: note Parquet I/O problems with Java 25+
It seems that somewhere between Java 21 and Java 25 (Java 23?)
the java platform effectively removed a method which is used by
the Hadoop libraries on which parquet-java relies.
This results in an UnspportedOperationException at runtime when
(at least some) parquet I/O operations are attempted; for instance
the unit tests fail.

At time of writing there's not much I can do about this,
so just note it in the user docs and advise to use an earlier
java version if you need Parquet support.