votable: handle char output more carefully
Be more careful about how character data is encoded to output.
Provide implementations for ASCII, UCS-2 and UTF-8.

Previous behaviour was something like UTF-8, though it called
itself ASCII.  For VOTable 1.6+, this will now write UTF-8.
For pre-VOTable 1.6 output the new behaviour differs somewhat from
what it did previously, but for reasons of pragmatism and continuity
it tries to do mostly the same thing with some corrections,
even though that can generate illegal VOTable output.

A switch is provided to make the pre-V1.6 behaviour more respectable;
this may be used in the future for instance when VOTable 1.6 output
becomes the default.