votable: update VOTable decoding to cope with UTF-8
The char and unicodeChar Decoder implementations are adjusted so
that they cope properly with UTF-8 and UTF-16 inputs respectively.
That is more lenient than what's required by VOTable, but it should
work with either the proposed VOTable 1.6 unicode handling or
the restricted content permitted by earlier versions, and do the
right thing for any legal content.  It will also work with certain
illegal forms of UTF-8/UTF-16 content.