votable: adjust broken XML character filtering
There is a method VOSerializer.ensureLegalXml which is supposed to
filter out characters which are simply illegal in XML, like \0.
But it was throwing out surrogate values of char types.

The source for this function was my copy of XML in a Nutshell,
which has the following production for an XML Char:

   Char ::= #x9 | #xA | #xD | [#x20-#xD7ff] | [#xE000-#xFFFD]
          | [#x10000-'x10FFFF ]

but my implementation of that was clearly falling foul of a poor
understanding of surrogate pairs as used in Java char types.
So I've removed the high boundaries, and now just throw out
values below x20 except for whitespace.