votable: fix UTF-8 bug in VOSerializer
Code was in place that would never have worked for writing
surrogate pairs to UTF-8-encoded XML e.g. in TABLEDATA output.
Fixing the problem as it stood would have been difficult as it
required manual encoding of characters to UTF-8, so don't try
to be clever, and instead leave it to the Charset implementation.
The first attempt at that slowed things down a lot, so rework
the implementation to write longer strings at a time for UTF-8
TABLEDATA output, which recovers most of the performance.

Although it's doing different things, the UTF-8 TABLEDATA performance
is more or less the same before and after this commit.