ttools: taplint allows reordering of UPL table
The taplint UPLoad stage did a SELECT * FROM tap_upload.xxx query,
expecting the returned table to be the same as the uploaded one.
This is too stringent, since DB tables are not ordered, so
the service may reorder the rows before returning it.
So add an ORDER BY clause to nail this down.

Reported by Jose Osinde at ESAC since the Greenplum DB they are
switching to does actually do such reordering.