votable: write INFOs and PARAMs in the right order
When writing out table parameters (DescribedValues) to VOTable,
they are mostly written as PARAMs but if they can't be for some
reason it falls back to INFOs.  The VOTable schema requires these
to be written in a fixed order (INFOs then PARAMs), they can't
be interspersed with each other at will, which was what was
happening before.  Fix this by accumulating all the INFOs and all
the PARAMs and outputting them once they're all available.

A similar remark applies to LINKs; in the current implementation
these can still get written in the wrong order, since they're
supposed to go after the FIELDs.  So this is a bug, but LINKs are
hardly ever written and it would be a pain to fix, so wait until
this crops up in practice before addressing it.