votable: write char not unicodeChar for v1.6 (usually)
When writing VOTable 1.6 VOTables or later, use the char datatype
(which is now UTF-8-capable) in preference to unicodeChar,
most of the time even if it's copying data that came from a VOTable
unicodeChar column.  We have to have an override for this to accommodate
stilts votcopy, which is supposed to retain all the VOTable document
structure apart from DATA serialization format.