votable: trust row count from data over metadata
The nrows attribute of a VOTable TABLE element should match the
row count reported by the TabularData it decorates, if both have
definite values.  In case of discrepancy, trust the data-derived
value rather than the metadata-derived one (and log a warning);
this is the opposite of the previous behaviour, but is less
likely to result in read errors.