topcat: fix subset update threading issue
Row Subset recounting was triggering Swing events from a thread
which was not the Event Dispatch Thread, which could cause
ConcurrentModificationException and possibly other problems
when lots of subset updates happened in quick succession,
e.g. as triggered by sliding a global variable control.

I've fixed this to dispatch such events asynchronously on the EDT.
I'm hoping it doesn't cause other problems, but it could do
so look out.