ttools: taplint F not E on 401/403
Taplint now tries to report a [F]ailure rather than an [E]rror
if it makes a query on a table that is inaccessible for reasons
of authentication or authorization; if it gets a 401 or 403
response it no longer signals it as an error.

I think this works for synchronous queries, though in the case of
async no 4xx codes come up, so those failures still look like errors.