vo: fix VOSI1.1 2-step metadata for delimited names
Table names which are delimited identifiers come wrapped in
double quotes (these are ubiquitous in TAPVizieR).
When they are reported thus by the /tables endpoint in the 2-step
VOSI metadata acquisition process, then the TAP window per-table
metadata acquisition was trying to make queries like:

   https://tapvizier.cds.unistra.fr/TAPVizieR/tap/tables/"II/246/out"

This caused a MalformedURLException (URISyntaxException) when passed
to URLUtils.newURL(), since a double quote (") is not a legal
URI character according to RFC3986.

It's not really clear what this URL should look like.  I'd guess:

   https://tapvizier.cds.unistra.fr/TAPVizieR/tap/tables/%22II/246/out%22

but maybe it should be:

   https://tapvizier.cds.unistra.fr/TAPVizieR/tap/tables/%22II%2F246%2Fout%22

When I identified this problem TAPVizieR wasn't responding usefully
to either, but Gilles Landais has fixed it so that the former works,
so fix the code up to work with that.  It's possible that other
services will have different ideas about whether to percent-encode
legal-but-reserved characters in the table name so watch this space,
but I'm not aware of any right now.