topcat: remove -out.meta=Dhul from Vizier API call
The parameter -out.meta=Dhul was being passed to the vizier query
service from topcat's Vizier load window.  It turns out that the
"l" flag in that value causes the result to return URLs instead
of the normal content of columns in some cases - e.g. the Cluster
column in J/A+A/640/A1 table1.  This is typically not desired
behaviour for download into topcat.

And according to the documentation, the default value for -out.meta
is "huD", so the default behaviour (no -out.meta parameter) is
what I want.  So just remove that parameter.

The behaviour of this service is documented, at time of writing, at
https://cdsarc.cds.unistra.fr/vizier/doc/vizquery.htx