vo: read column and FK TAP metadata serially
Modify the code in TableSetPanel so that when a table is selected
the column metadata is read first, and then the foreign key metadata;
previously they were read concurrently.

The reason is that for some metadata acquisition modes (VOSI1.1-2step)
you need to make an identical query to get column and FK information,
so there is the possibility to make only one of those queries,
which will automatically update the other, rather than making
two identical queries at the same time.

At this commit, two identical queries are still made (sequentially).