vo: VOSI column queries update FKs and vice versa
Fix the Vosi11TapMetaReader readColumns and radForeignKeys methods
so that each updates both columns and FKs - each requires making
the same query to the tables endpoint, and gets enough information
to do both things.  Along with a recent change to TableSetPanel this
means we can make only a single query rather than two identical
queries to the tables endpoint if both column and FK metadata
are required.

The changes here are a bit questionable, since they operate via
side-effects that mutate the TableMeta object which is really
intended as a read-only input value to designate the identity of
the table to query for, but which is actually the same object that
gets mutated according to the result of these methods.
So there is possibility for logic confusion here, but as currently
used I believe it does the right thing.

Backing out of this commit would not break anything, but it would
revert to the state in which two identical queries got done one
after the other in VOSI1.1-2step mode.