topcat: fix issue with autocomplete popups
The popup autocomplete menus for column values were not registering
a selection if you just clicked on one of the items in the popup
without subsequently hitting return in the corresponding text field.
This meant you could see what looked like the right items in the
fields, but no plot showed up.  It was a bug in managing the
underlying ComboBoxModel - fix it.

The autocomplete was introduced since the last public release,
so this bug was never visible in the public release.